Move BinaryQuadraticModel under the dimod.bqm namespace
