+++ dimod/bqm/__init__.py
@@ -1,3 +1,5 @@
 from dimod.bqm.vectorbqm import VectorBinaryQuadraticModel, VectorBQM
 from dimod.bqm.fastbqm import FastBinaryQuadraticModel, FastBQM
 from dimod.bqm.utils import sum
+
+from dimod.bqm.binary_quadratic_model import BinaryQuadraticModel, OrderedBinaryQuadraticModel
