+++ dimod/bqm/binary_quadratic_model.py
