+++ dimod/generators/chimera.py
@@ -17,7 +17,7 @@ from __future__ import absolute_import
 
 import random
 
-from dimod.binary_quadratic_model import BinaryQuadraticModel
+from dimod.bqm.binary_quadratic_model import BinaryQuadraticModel
 from dimod.vartypes import SPIN
 
 __all__ = ['chimera_anticluster']
