+++ dimod/io/json.py
@@ -28,7 +28,7 @@ import numpy as np
 
 from six import iteritems
 
-from dimod.binary_quadratic_model import BinaryQuadraticModel
+from dimod.bqm.binary_quadratic_model import BinaryQuadraticModel
 from dimod.package_info import __version__
 from dimod.sampleset import SampleSet
 from dimod.vartypes import Vartype
